refactor: delete verseUtils
